Fixes for types

+++ eppoclient/client.go
@@ -122,5 +122,5 @@ func isInExperimentSample(subjectKey string, experimentKey string, experimentCon
 	shardKey := "exposure-" + subjectKey + "-" + experimentKey
 	shard := getShard(shardKey, int64(experimentConfig.SubjectShards))
 
-	return shard <= int64(experimentConfig.PercentExposure)*int64(experimentConfig.SubjectShards)
+	return float64(shard) <= float64(experimentConfig.PercentExposure)*float64(experimentConfig.SubjectShards)
 }
